refactor: rename file for autocomplete component to reflect comp name
