+++ libs/spa/core/ui/src/lib/entity-search/component/autocomplete.directive.ts
